+++ modules/gravitino/server/versions.tf
@@ -1,18 +1,19 @@
 terraform {
-  required_version = ">= 1.0"
+  required_version = ">= 1.14.3"
 
   required_providers {
     kubernetes = {
       source  = "hashicorp/kubernetes"
-      version = "~> 2.30.0"
+      version = "~> 3.0.1"
+
     }
     helm = {
       source  = "hashicorp/helm"
       version = "~> 3.1.1"
     }
     external = {
-      source  = "hashicorp/external"
-      version = "~> 2.0"
+      source = "hashicorp/external"
+
     }
   }
 }